Dosyalar gerekli katmanlara ayrildi. Baglantilar saglandi. Controller siniflari iliskisi saglandi.
