+++ lib/Controller/InputController.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/cupertino.dart';
+
+class InputController{
+  TextEditingController nameController = TextEditingController();
+  TextEditingController surnameController = TextEditingController();
+  TextEditingController phoneNumberController = TextEditingController();
+  TextEditingController shuttleCodeController = TextEditingController();
+}
