--- lib/Entities/AbstractUser.dart
@@ -1,7 +1,0 @@
-
-
-abstract class AbstractUser{
-  String? name;
-  String? surname;
-  String? phone_number;
-}
