--- lib/Entities/Children.dart
@@ -1,5 +1,0 @@
-class Children{
-  String? parent_phone;
-  String? school_id;
-  String? service_code;
-}
