+++ lib/Entities/Concretes/Shuttle.dart
@@ -1,0 +1,11 @@
+import 'package:mobile_dev/Entities/Concretes/Driver.dart';
+import 'package:mobile_dev/Pages/register_hostes.dart';
+
+import 'Hostess.dart';
+
+class Shuttle{
+  String? plate_number;
+  String? school_id;
+  late Driver driver;
+  late Hostess hostess;
+}
