--- lib/Entities/Driver.dart
@@ -1,5 +1,0 @@
-import 'package:mobile_dev/Entities/AbstractUser.dart';
-
-class Driver extends AbstractUser{
-
-}
