--- lib/Entities/Hostess.dart
@@ -1,6 +1,0 @@
-import 'package:mobile_dev/Entities/AbstractUser.dart';
-import 'package:mobile_dev/Entities/Children.dart';
-
-class Hostess extends AbstractUser{
-  late List<Children> child_list;
-}
