--- lib/Entities/Parent.dart
@@ -1,9 +1,0 @@
-import 'package:mobile_dev/Entities/AbstractUser.dart';
-
-import 'Children.dart';
-
-class Parent extends AbstractUser{
-  String?  phoneNumber;
-  String? shuttle_key;
-  late List<Children> childList;
-}
