--- lib/Entities/School.dart
@@ -1,7 +1,0 @@
-import 'Shuttle.dart';
-
-class School{
-  String? school_name;
-  String? school_id;
-  late List<Shuttle> shuttle_list;
-}
