--- lib/Entities/Shuttle.dart
@@ -1,9 +1,0 @@
-import 'package:mobile_dev/Entities/Driver.dart';
-import 'package:mobile_dev/register_hostes.dart';
-
-class Shuttle{
-  String? plate_number;
-  String? school_id;
-  late Driver driver;
-  late Hostess hostess;
-}
