+++ lib/Pages/register_parent.dart
@@ -1,0 +1,90 @@
+// add_child.dart
+import 'package:mobile_dev/Controller/InputController.dart';
+import 'package:mobile_dev/Controller/ParentRegisterController.dart';
+import 'package:mobile_dev/Entities/Concretes/Parent.dart';
+
+import '../main.dart';
+
+import 'package:flutter/material.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+
+
+class ParentRegister extends StatelessWidget {
+
+  // Boş bir dropdown değeri
+  Parent parent=Parent();
+  InputController inputController=InputController();
+  ParentRegisterController parentRegisterController=ParentRegisterController();
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(
+        primarySwatch: Colors.red,
+      ),
+      home: Scaffold(
+        body: Center(
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text(
+                  "Parent-Sign Up",
+                  style: TextStyle(
+                    fontSize: 24.0,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                TextField(
+                  controller: inputController.nameController,
+                  decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: "Enter name",
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                TextField(
+                  controller: inputController.surnameController,
+                  decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: "Surname",
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                TextField(
+                  controller: inputController.phoneNumberController,
+                  decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: "Enter Phone Number",
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                TextField(
+                  obscureText: true,
+                  controller: inputController.shuttleCodeController,
+                  decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: "Shuttle Code",
+                  ),
+                ),
+                SizedBox(height: 16.0),
+                ElevatedButton(
+                  onPressed: () async {
+                    parentRegisterController.register(parent,inputController);
+                    // Additional logic or navigation can be added here
+                  },
+                  child: Text("Register"),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
